backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0  # MySQL-Python connector
alembic==1.12.1

# Machine Learning
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
xgboost==2.0.1
lime==0.2.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0.post1
python-jose==3.3.0  # For JWT
passlib==1.7.4  # For password hashing

# Note: Kafka dependencies have been removed as they are not needed for local development
